Deprecate old parsers and add examples for updates
