+++ test/data/raw_outputs/amrfinder/report.tsv
@@ -1,0 +1,20 @@
+Protein identifier	Contig id	Start	Stop	Strand	Gene symbol	Sequence name	Scope	Element type	Element subtype	Class	Subclass	Method	Target length	Reference sequence length	% Coverage of reference sequence	% Identity to reference sequence	Alignment length	Accession of closest sequence	Name of closest sequence	HMM id	HMM description
+NA	NODE_11_length_157792_cov_39.2115	128890	129747	+	blaSHV-11	class A broad-spectrum beta-lactamase SHV-11	core	AMR	AMR	BETA-LACTAM	BETA-LACTAM	ALLELEX	286	286	100.00	100.00	286	WP_004176269.1	class A broad-spectrum beta-lactamase SHV-11	NF000285.3	SHV family class A beta-lactamase
+NA	NODE_123_length_431_cov_19.8947	234	431	+	qacE	quaternary ammonium compound efflux SMR transporter QacE	core	STRESS	BIOCIDE	QUATERNARY AMMONIUM	QUATERNARY AMMONIUM	PARTIAL_CONTIG_ENDX	66	115	57.39	100.00	66	WP_000679427.1	quaternary ammonium compound efflux SMR transporter QacE delta 1	NF000276.2	quaternary ammonium compound efflux SMR transporter QacE
+NA	NODE_1_length_803036_cov_42.265	30758	31930	+	oqxA	multidrug efflux RND transporter periplasmic adaptor subunit OqxA	core	AMR	AMR	PHENICOL/QUINOLONE	PHENICOL/QUINOLONE	EXACTX	391	391	100.00	100.00	391	WP_002914189.1	multidrug efflux RND transporter periplasmic adaptor subunit OqxA	NF000272.1	multidrug efflux RND transporter periplasmic adaptor subunit OqxA
+NA	NODE_1_length_803036_cov_42.265	31957	35106	+	oqxB	multidrug efflux RND transporter permease subunit OqxB	core	AMR	AMR	PHENICOL/QUINOLONE	PHENICOL/QUINOLONE	BLASTX	1050	1050	100.00	99.90	1050	WP_021312368.1	multidrug efflux RND transporter permease subunit OqxB20	NF000037.1	multidrug efflux RND transporter permease subunit OqxB
+NA	NODE_39_length_11164_cov_23.0187	4569	5426	+	aac(3)-IId	aminoglycoside N-acetyltransferase AAC(3)-IId	core	AMR	AMR	AMINOGLYCOSIDE	GENTAMICIN	EXACTX	286	286	100.00	100.00	286	WP_000557454.1	aminoglycoside N-acetyltransferase AAC(3)-IId	NF033080.0	AAC(3)-II family aminoglycoside N-acetyltransferase
+NA	NODE_3_length_454317_cov_40.2449	161022	163652	-	gyrA_D87A	Klebsiella quinolone resistant GyrA	core	AMR	POINT	QUINOLONE	QUINOLONE	POINTX	877	877	100.00	99.66	877	WP_117036963.1	DNA gyrase subunit A GyrA	NA	NA
+NA	NODE_3_length_454317_cov_40.2449	161022	163652	-	gyrA_S83F	Klebsiella quinolone resistant GyrA	core	AMR	POINT	QUINOLONE	QUINOLONE	POINTX	877	877	100.00	99.66	877	WP_117036963.1	DNA gyrase subunit A GyrA	NA	NA
+NA	NODE_40_length_7360_cov_24.8738	6361	7203	-	rmtC	RmtC family 16S rRNA (guanine(1405)-N(7))-methyltransferase	core	AMR	AMR	AMINOGLYCOSIDE	AMINOGLYCOSIDE	EXACTX	281	281	100.00	100.00	281	WP_000855769.1	RmtC family 16S rRNA (guanine(1405)-N(7))-methyltransferase	NF000150.1	RmtC family 16S rRNA (guanine(1405)-N(7))-methyltransferase
+NA	NODE_44_length_4295_cov_23.2123	301	1110	+	blaNDM-1	subclass B1 metallo-beta-lactamase NDM-1	core	AMR	AMR	BETA-LACTAM	CARBAPENEM	ALLELEX	270	270	100.00	100.00	270	WP_004201164.1	subclass B1 metallo-beta-lactamase NDM-1	NF000259.2	NDM family subclass B1 metallo-beta-lactamase
+NA	NODE_44_length_4295_cov_23.2123	1117	1479	+	ble	bleomycin binding protein Ble-MBL	core	AMR	AMR	BLEOMYCIN	BLEOMYCIN	EXACTX	121	121	100.00	100.00	121	WP_004201167.1	bleomycin binding protein Ble-MBL	NF012202.0	bleomycin binding protein Ble-MBL
+NA	NODE_44_length_4295_cov_23.2123	2068	2862	-	blaDHA	DHA family class C beta-lactamase	core	AMR	AMR	BETA-LACTAM	CEPHALOSPORIN	PARTIALX	265	379	69.92	100.00	265	WP_004236386.1	class C beta-lactamase DHA-1	NF012102.1	DHA family class C beta-lactamase
+NA	NODE_47_length_3813_cov_27.9457	2766	3638	-	blaCTX-M-15	class A extended-spectrum beta-lactamase CTX-M-15	core	AMR	AMR	BETA-LACTAM	CEPHALOSPORIN	ALLELEX	291	291	100.00	100.00	291	WP_000239590.1	class A extended-spectrum beta-lactamase CTX-M-15	NF033089.1	CTX-M family class A extended-spectrum beta-lactamase
+NA	NODE_4_length_367278_cov_42.3139	14037	14453	+	fosA	FosA family fosfomycin resistance glutathione transferase	core	AMR	AMR	FOSFOMYCIN	FOSFOMYCIN	BLASTX	139	139	100.00	98.56	139	WP_004146118.1	FosA family fosfomycin resistance glutathione transferase	NF000221.4	FosA family fosfomycin resistance glutathione transferase
+NA	NODE_51_length_1731_cov_22.5767	196	1023	+	blaOXA-1	oxacillin-hydrolyzing class D beta-lactamase OXA-1	core	AMR	AMR	BETA-LACTAM	BETA-LACTAM	ALLELEX	276	276	100.00	100.00	276	WP_001334766.1	oxacillin-hydrolyzing class D beta-lactamase OXA-1	NF000388.2	OXA-1 family class D beta-lactamase
+NA	NODE_51_length_1731_cov_22.5767	1164	1604	+	catB3	type B-3 chloramphenicol O-acetyltransferase CatB3	core	AMR	AMR	PHENICOL	CHLORAMPHENICOL	PARTIALX	147	210	70.00	100.00	147	WP_000186237.1	type B-3 chloramphenicol O-acetyltransferase CatB3	NF000490.1	type B chloramphenicol O-acetyltransferase
+NA	NODE_57_length_1134_cov_34.8908	28	864	-	sul1	sulfonamide-resistant dihydropteroate synthase Sul1	core	AMR	AMR	SULFONAMIDE	SULFONAMIDE	EXACTX	279	279	100.00	100.00	279	WP_000259031.1	sulfonamide-resistant dihydropteroate synthase Sul1	NF000294.1	sulfonamide-resistant dihydropteroate synthase Sul1
+NA	NODE_57_length_1134_cov_34.8908	861	1133	-	qacE	quaternary ammonium compound efflux SMR transporter QacE	core	STRESS	BIOCIDE	QUATERNARY AMMONIUM	QUATERNARY AMMONIUM	PARTIAL_CONTIG_ENDX	91	115	79.13	100.00	91	WP_000679427.1	quaternary ammonium compound efflux SMR transporter QacE delta 1	NF000276.2	quaternary ammonium compound efflux SMR transporter QacE
+NA	NODE_69_length_625_cov_55.4398	66	617	-	aac(6')-Ib3	aminoglycoside N-acetyltransferase AAC(6')-Ib3	core	AMR	AMR	AMINOGLYCOSIDE	AMIKACIN/KANAMYCIN/TOBRAMYCIN	ALLELEX	184	184	100.00	100.00	184	WP_032488579.1	aminoglycoside N-acetyltransferase AAC(6')-Ib3	NF033074.0	AAC(6')-Ib family aminoglycoside 6'-N-acetyltransferase
+NA	NODE_7_length_251151_cov_40.8784	231802	232896	-	pmrB_R256G	Klebsiella colistin resistant PmrB	core	AMR	POINT	COLISTIN	COLISTIN	POINTX	365	365	100.00	99.45	365	WP_004179093.1	two-component system sensor histidine kinase PmrB	NA	NA
