+++ test/data/raw_outputs/groot/report.tsv
@@ -1,0 +1,22 @@
+efpA.3003955.AL123456.3.3153038-3154631.5259	82	1593	1582M11D
+Mycobacterium_tuberculosis_16S.3003481.AL123456.3.1471846-1473382.3261	86	1537	8D1521M8D
+murA.3003784.HE608151.1472985-1474242.3555	96	1257	3D1238M16D
+Mycobacterium_tuberculosis_16S.3003436.AL123456.3.1471846-1473382.3263	86	1537	8D1521M8D
+Mycobacterium_tuberculosis_16S.3003437.AL123456.3.1471846-1473382.3260	86	1537	8D1521M8D
+iniC.3003451.AL123456.412756-414238.2100	106	1482	2D1450M30D
+embR.3003455.AL123456.3.1416180-1417347.4712	91	1167	13D614M1D519M20D
+Mycobacterium_tuberculosis_16S.3003480.AL123456.3.1471845-1473382.4123	86	1537	8D1521M8D
+katG.3003392.AL123456.3.2153888-2156111.5483	144	2223	1D2216M6D
+rpsA.3004721.AL123456.3.1833541-1834987.5598	78	1446	10D538M4D885M9D
+embB.3003325.BX248333.4182808-4186105.3355	223	3297	40D310M39D743M9D2156M
+rpoB.3003283.AL123456.3.759806-763325.4199	246	3519	5D1402M10D1076M10D1000M16D
+ethA.3003458.AE000516.2.4318327-4319797.4455	100	1470	20D1445M5D
+embB.3003465.AE000516.4238826-4242123.3327	227	3297	7D2869M32D389M
+gyrA.3003295.AL123456.7301-9818.2055	183	2517	10D2492M15D
+mfpA.3003035.AL123456.3773015-3773567.4588	36	552	1D537M14D
+gidB.3003470.AL123456.4407527-4408202.4553	43	675	5D665M5D
+folC.3004157.AL123456.3.2746134-2747598.5391	96	1464	7D1444M13D
+embC.3003327.CP003248.4240040-4243325.2080	260	3285	23D3259M3D
+embA.3003453.CP003248.4243410-4246695.2081	198	3285	2D148M2D1243M1D1878M11D
+embB.3003326.AL123456.4246513-4249810.2078	236	3297	7D3290M
+iniA.3003448.AL123456.410837-412760.2099	138	1923	20D1898M5D
