+++ test/data/raw_outputs/kmerresistance/results.res
@@ -1,0 +1,62 @@
+#Template	Score	Expected	Template_length	Template_Identity	Template_Coverage	Query_Identity	Query_Coverage	Depth	q_value	p_value
+aac(6')-Ib11_1_AY136758	     888	     196	     570	   52.98	   56.67	   93.50	  176.47	    1.57	  441.51	1.0e-26
+aac(6')-Ib-Hangzhou_1_FJ503047	     231	     178	     519	   34.10	   38.34	   88.94	  260.80	    0.76	    6.67	9.8e-03
+aac(6')-Ib-Suzhou_1_EU085533	    2697	     178	     519	   45.09	   46.05	   97.91	  217.15	    6.61	 2206.41	1.0e-26
+aac(3)-IId_1_EU022314	   50687	     281	     861	   99.88	  100.00	   99.88	  100.00	   65.40	49849.41	1.0e-26
+aac(6')-Ib_1_M21682	     581	     208	     606	   37.62	   40.26	   93.44	  248.36	    1.60	  175.63	1.0e-26
+aac(6')-Ib3_1_X60321	   33512	     184	     555	  100.00	  100.00	  100.00	  100.00	   65.06	32962.16	1.0e-26
+ant(3'')-Ia_1_X02340	    5885	     332	     972	   20.06	   24.18	   82.98	  413.62	    7.65	 4957.78	1.0e-26
+rmtC_1_AB194779	   54314	     275	     846	  100.00	  100.00	  100.00	  100.00	   71.18	53493.69	1.0e-26
+aac(6')-Ib-cr_1_DQ303918	   46759	     196	     600	  100.00	  100.00	  100.00	  100.00	   87.95	46171.86	1.0e-26
+aac(6')-Ib-cr_2_EF636461	     248	     178	     519	   48.36	   48.36	  100.00	  206.77	    0.48	   11.25	8.0e-04
+blaNDM-18_1_KY503030	     407	     285	     828	   31.88	   36.11	   88.29	  276.92	    0.61	   21.47	3.6e-06
+blaNDM-16_1_KP862821	     567	     279	     813	   27.92	   30.26	   92.28	  330.49	    1.00	   97.34	1.0e-22
+blaCTX-M-54_1_DQ303459	     360	     301	     876	   26.03	   30.71	   84.76	  325.65	    0.59	    5.15	2.3e-02
+blaCTX-M-64_1_AB284167	     440	     301	     876	   26.60	   32.88	   80.90	  304.17	    0.90	   25.82	3.8e-07
+blaCTX-M-101_1_HQ398214	     657	     301	     876	   49.66	   53.77	   92.36	  185.99	    1.17	  131.81	1.0e-26
+blaDHA-1_1_Y16410	   38868	     377	    1140	   73.07	   76.32	   95.75	  131.03	   38.13	37750.98	1.0e-26
+blaDHA-3_1_AY494945	     744	     392	    1140	   24.39	   27.98	   87.15	  357.37	    0.89	  108.76	1.0e-24
+blaDHA-5_1_JF273491	    1347	     392	    1140	   58.07	   69.39	   83.69	  144.12	    1.87	  524.19	1.0e-26
+blaOXA-1_1_HQ170510	   36876	     275	     831	  100.00	  100.00	  100.00	  100.00	   48.93	36057.70	1.0e-26
+blaSHV-36_1_AF467947	     440	     296	     861	   22.53	   24.51	   91.94	  408.06	    0.66	   27.98	1.2e-07
+blaSHV-51_1_AY289548	     612	     296	     861	   43.09	   44.60	   96.61	  224.22	    0.87	  109.65	1.0e-24
+blaSHV-61_1_AJ866284	     918	     296	     861	   25.32	   26.02	   97.32	  384.38	    1.22	  318.29	1.0e-26
+blaSHV-64_1_DQ174304	    1552	     296	     861	   57.14	   64.23	   88.97	  155.70	    2.61	  853.43	1.0e-26
+blaSHV-67_1_DQ174307	     513	     296	     861	   29.73	   33.91	   87.67	  294.86	    0.89	   57.95	1.0e-13
+blaSHV-74_1_AM176549	     634	     296	     861	   28.11	   30.43	   92.37	  328.63	    1.19	  122.51	1.0e-26
+blaSHV-76_1_AM176551	     409	     296	     861	   28.46	   30.89	   92.11	  323.68	    0.59	   17.95	2.3e-05
+blaSHV-93_1_EF373969	     369	     296	     861	   29.15	   31.13	   93.66	  321.27	    0.56	    7.90	4.9e-03
+blaSHV-94_1_EF373970	     856	     296	     861	   58.30	   61.56	   94.72	  162.45	    1.18	  271.83	1.0e-26
+blaSHV-99_1_AM941845	     446	     296	     861	   28.57	   35.89	   79.61	  278.64	    0.94	   30.12	4.1e-08
+blaSHV-155_1_JX121122	     552	     295	     858	   29.25	   29.84	   98.05	  335.16	    0.78	   77.71	1.0e-17
+blaSHV-156_1_JX121123	     397	     296	     861	   27.99	   31.48	   88.93	  317.71	    0.60	   14.57	1.3e-04
+blaSHV-157_1_JX121124	     387	     296	     861	   25.90	   27.99	   92.53	  357.26	    0.55	   11.99	5.3e-04
+blaSHV-159_1_JX121126	     541	     295	     858	   22.03	   22.14	   99.47	  451.58	    0.87	   72.13	1.0e-16
+blaNDM-1_1_FN396876	   33696	     270	     813	  100.00	  100.00	  100.00	  100.00	   44.34	32893.32	1.0e-26
+blaNDM-24_1_MH450215	     370	     279	     813	   32.72	   37.64	   86.93	  265.69	    0.62	   12.48	4.1e-04
+blaCTX-M-163_1_KP681698	     437	     301	     876	   29.00	   29.91	   96.95	  334.35	    0.55	   24.81	6.3e-07
+blaCTX-M-167_1_KR537428	     505	     301	     876	   28.08	   31.16	   90.11	  320.88	    0.86	   51.29	1.0e-12
+blaCTX-M-177_1_KT997889	     523	     301	     876	   44.52	   47.37	   93.98	  211.08	    0.80	   59.45	1.0e-13
+blaCTX-M-184_1_KX263247	     691	     301	     876	   42.81	   48.06	   89.07	  208.08	    1.21	  152.82	1.0e-26
+blaCTX-M-204_1_MF281085	     501	     301	     876	   37.33	   37.33	  100.00	  267.89	    0.57	   49.54	1.0e-11
+blaCTX-M-209_1_MH067959	     408	     301	     876	   28.31	   30.25	   93.58	  330.57	    0.58	   15.94	6.5e-05
+blaCTX-M-15_1_AY044436	   43643	     288	     876	  100.00	  100.00	  100.00	  100.00	   51.90	42785.57	1.0e-26
+blaDHA-25_1_KY563770	     497	     392	    1140	   22.37	   24.82	   90.11	  402.83	    0.68	   12.27	4.6e-04
+blaDHA-14_1_DQ478687	     682	     392	    1140	   25.44	   28.86	   88.15	  346.50	    0.87	   78.03	1.0e-17
+blaSHV-11_1_X98101	   16127	     291	     861	   56.91	   56.91	  100.00	  175.71	   20.35	15272.60	1.0e-26
+blaSHV-119_1_JX192924	     522	     296	     861	   31.13	   33.45	   93.06	  298.96	    0.80	   62.17	1.0e-14
+blaSHV-182_1_KP050489	   49808	     281	     861	   99.42	  100.81	   98.62	   99.19	   60.14	48969.73	1.0e-26
+fosA5_1_EU195449	    3004	     144	     420	   52.38	   54.76	   95.65	  182.61	    8.53	 2597.79	1.0e-26
+fosA_3_ACWO01000079	   24768	     141	     420	   98.57	  100.00	   98.57	  100.00	   65.36	24348.19	1.0e-26
+fosA_6_ACZD01000244	     897	     144	     420	   21.90	   22.62	   96.84	  442.11	    2.35	  543.65	1.0e-26
+mdf(A)_1_Y08743	    4466	     422	    1233	   27.66	   33.33	   82.97	  300.00	    6.91	 3343.75	1.0e-26
+catB3_1_AJ009818	   24169	     212	     633	   78.20	   84.20	   92.87	  118.76	   43.24	23538.44	1.0e-26
+oqxA_1_EU370913	  110587	     359	    1176	   99.23	  100.00	   99.23	  100.00	  105.36	109513.22	1.0e-26
+oqxB_1_EU370913	  293480	     761	    3153	   98.86	  100.00	   98.86	  100.00	  104.05	291202.21	1.0e-26
+sul1_2_U12338	   52742	     273	     840	   99.88	  100.12	   99.76	   99.88	   66.35	51926.26	1.0e-26
+sul1_3_EU855787	     439	     289	     840	   30.48	   32.02	   95.17	  312.27	    0.60	   30.81	2.9e-08
+sul1_5_EU780013	   14322	     294	     867	   60.55	   60.67	   99.81	  164.83	   17.27	13462.82	1.0e-26
+sul1_10_DQ143913	    1645	     285	     831	   64.38	   73.41	   87.70	  136.23	    3.03	  956.88	1.0e-26
+sul1_30_JF262178	     724	     289	     840	   48.10	   51.07	   94.17	  195.80	    1.19	  186.65	1.0e-26
+sul1_39_AY522923	     778	     284	     828	   71.62	   76.57	   93.53	  130.60	    1.36	  228.66	1.0e-26
+sul1_37_AF188331	     502	     289	     840	   75.95	   85.12	   89.23	  117.48	    0.90	   57.23	1.0e-13
