+++ test/data/raw_outputs/mykrobe/report.json
@@ -1,0 +1,117 @@
+{
+    "my_sample_name": {
+        "susceptibility": {
+            "Ofloxacin": {
+                "predict": "S"
+            },
+            "Moxifloxacin": {
+                "predict": "S"
+            },
+            "Isoniazid": {
+                "predict": "R",
+                "called_by": {
+                    "inhA_I21T-ATC1674262ACT": {
+                        "variant": null,
+                        "genotype": [
+                            1,
+                            1
+                        ],
+                        "genotype_likelihoods": [
+                            -467.0296062525409,
+                            -99999999,
+                            -8.903844781805352
+                        ],
+                        "info": {
+                            "coverage": {
+                                "reference": {
+                                    "percent_coverage": 0.0,
+                                    "median_depth": 0,
+                                    "min_non_zero_depth": 0,
+                                    "kmer_count": 0,
+                                    "klen": 21
+                                },
+                                "alternate": {
+                                    "percent_coverage": 100.0,
+                                    "median_depth": 4,
+                                    "min_non_zero_depth": 3,
+                                    "kmer_count": 72,
+                                    "klen": 20
+                                }
+                            },
+                            "expected_depths": [
+                                4
+                            ],
+                            "contamination_depths": [],
+                            "filter": [],
+                            "conf": 458
+                        },
+                        "_cls": "Call.VariantCall"
+                    }
+                }
+            },
+            "Kanamycin": {
+                "predict": "S"
+            },
+            "Ethambutol": {
+                "predict": "S"
+            },
+            "Streptomycin": {
+                "predict": "S"
+            },
+            "Ciprofloxacin": {
+                "predict": "S"
+            },
+            "Pyrazinamide": {
+                "predict": "S"
+            },
+            "Rifampicin": {
+                "predict": "S"
+            },
+            "Amikacin": {
+                "predict": "S"
+            },
+            "Capreomycin": {
+                "predict": "S"
+            }
+        },
+        "phylogenetics": {
+            "phylo_group": {
+                "Mycobacterium_tuberculosis_complex": {
+                    "percent_coverage": 95.435,
+                    "median_depth": 4
+                }
+            },
+            "sub_complex": {
+                "Unknown": {
+                    "percent_coverage": -1,
+                    "median_depth": -1
+                }
+            },
+            "species": {
+                "Mycobacterium_tuberculosis": {
+                    "percent_coverage": 94.177,
+                    "median_depth": 4.0
+                }
+            },
+            "lineage": {
+                "European_American": {
+                    "percent_coverage": 65.833,
+                    "median_depth": 3.5
+                }
+            }
+        },
+        "kmer": 21,
+        "probe_sets": [
+            "/home/fin/miniconda3/envs/amrtest/lib/python3.7/site-packages/mykrobe/data/panels/tb-species-170421.fasta.gz",
+            "/home/fin/miniconda3/envs/amrtest/lib/python3.7/site-packages/mykrobe/data/panels/tb-hunt-probe-set-jan-03-2019.fasta.gz"
+        ],
+        "files": [
+            "reads.fq.gz"
+        ],
+        "version": {
+            "mykrobe-predictor": "v0.8.1",
+            "mykrobe-atlas": "v0.8.1"
+        },
+        "genotype_model": "kmer_count"
+    }
+}
