+++ test/data/raw_outputs/pointfinder/report.tsv
@@ -1,0 +1,21 @@
+Mutation	Nucleotide change	Amino acid change	Resistance	PMID
+acrR p.P161R	CCG -> CGG	P -> R	fluoroquinolone	12936981
+acrR p.G164A	GGC -> GCC	G -> A	fluoroquinolone	12936981
+acrR p.F172S	TTC -> TCC	F -> S	fluoroquinolone	12936981
+acrR p.R173G	CGA -> GGG	R -> G	fluoroquinolone	12936981
+acrR p.L195V	CTC -> GTC	L -> V	fluoroquinolone	12936981
+acrR p.F197I	TTT -> ATT	F -> I	fluoroquinolone	12936981
+acrR p.K201M	AAG -> ATG	K -> M	fluoroquinolone	12936981
+ompK36 p.N49S	AAC -> AGC	N -> S	cephalosporins	25245001
+ompK36 p.L59V	CTT -> GTA	L -> V	cephalosporins	25245001
+ompK36 p.G189T	GGC -> ACC	G -> T	cephalosporins	25245001
+ompK36 p.F198Y	TTC -> TAC	F -> Y	cephalosporins	25245001
+ompK36 p.F207Y	TTC -> TAT	F -> Y	cephalosporins	25245001
+ompK36 p.A217S	GCT -> TCT	A -> S	carbapenem	25245001
+ompK36 p.T222L	ACC -> CTT	T -> L	cephalosporins	25245001
+ompK36 p.D223G	GAC -> GGC	D -> G	cephalosporins	25245001
+ompK36 p.E232R	GAA -> CGT	E -> R	cephalosporins	25245001
+ompK36 p.N304E	AAC -> GAA	N -> E	cephalosporins	25245001
+ompK37 p.I70M	ATT -> ATG	I -> M	carbapenem	22282462
+ompK37 p.I128M	ATT -> ATG	I -> M	carbapenem	22282462
+ompK37 p.N230G	AAC -> GGC	N -> G	carbapenem	22282462
