+++ test/data/raw_outputs/sstar/report.tsv
@@ -1,0 +1,18 @@
+blaSHV*	blaSHV-100*	NODE_11_length_157792_cov_39.2115	95.111%	900	900
+catB3?$	catB3?$	NODE_51_length_1731_cov_22.5767	99.775%	445	633
+ampH*	ampH*	NODE_6_length_262587_cov_42.4347	98.88%	1161	1161
+aac(6')?$	aac(6')-30-aac(6')-Ib'?$	NODE_69_length_625_cov_55.4398	99.645%	563	987
+aac(3)*	aac(3)-IId*	NODE_39_length_11164_cov_23.0187	99.884%	861	861
+sul1*	sul1*	NODE_57_length_1134_cov_34.8908	99.769%	867	867
+blaOXA	blaOXA-1	NODE_51_length_1731_cov_22.5767	100.0%	831	831
+oqxA*	oqxA*	NODE_1_length_803036_cov_42.265	99.235%	1176	1176
+blaSHV	blaSHV-11	NODE_11_length_157792_cov_39.2115	100.0%	861	861
+catB4	catB4	NODE_51_length_1731_cov_22.5767	100.0%	549	549
+oqxB*	oqxB*	NODE_1_length_803036_cov_42.265	98.858%	3153	3153
+blaCTX	blaCTX-M-15	NODE_47_length_3813_cov_27.9457	100.0%	876	876
+fosA6?$	fosA6?$	NODE_4_length_367278_cov_42.3139	98.81%	420	433
+aac(3)?$	aac(3)-Ib-aac(6')-Ib'?$	NODE_69_length_625_cov_55.4398	99.458%	554	1005
+blaNDM	blaNDM-1	NODE_44_length_4295_cov_23.2123	100.0%	813	813
+aac(6')?	aac(6')-Ib?	NODE_69_length_625_cov_55.4398	99.822%	563	606
+blaDHA?$	blaDHA-1?$	NODE_44_length_4295_cov_23.2123	100.0%	796	1140
+rmtC	rmtC	NODE_40_length_7360_cov_24.8738	100.0%	846	846
